gh-116738: Move run_concurrently() to threading_helper
